# [BEAM-13015, #22050] Make SDK harness msec counters faster using ordered puts