--- a/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/control/ProcessBundleHandler.java
+++ b/sdks/java/harness/src/main/java/org/apache/beam/fn/harness/control/ProcessBundleHandler.java
@@ -18,7 +18,6 @@
 package org.apache.beam.fn.harness.control;
 
 import com.google.auto.value.AutoValue;
-import java.io.Closeable;
 import java.io.IOException;
 import java.time.Duration;
 import java.util.ArrayList;
@@ -48,6 +47,7 @@ import org.apache.beam.fn.harness.Caches.ClearableCache;
 import org.apache.beam.fn.harness.PTransformRunnerFactory;
 import org.apache.beam.fn.harness.PTransformRunnerFactory.Context;
 import org.apache.beam.fn.harness.PTransformRunnerFactory.Registrar;
+import org.apache.beam.fn.harness.control.ExecutionStateSampler.ExecutionStateTracker;
 import org.apache.beam.fn.harness.control.FinalizeBundleHandler.CallbackRegistration;
 import org.apache.beam.fn.harness.data.BeamFnDataClient;
 import org.apache.beam.fn.harness.data.PCollectionConsumerRegistry;
@@ -74,9 +74,8 @@ import org.apache.beam.model.pipeline.v1.RunnerApi.WindowingStrategy;
 import org.apache.beam.runners.core.construction.BeamUrns;
 import org.apache.beam.runners.core.construction.PTransformTranslation;
 import org.apache.beam.runners.core.construction.Timer;
-import org.apache.beam.runners.core.metrics.ExecutionStateSampler;
-import org.apache.beam.runners.core.metrics.ExecutionStateTracker;
 import org.apache.beam.runners.core.metrics.MetricsContainerStepMap;
+import org.apache.beam.runners.core.metrics.MonitoringInfoConstants.Urns;
 import org.apache.beam.runners.core.metrics.ShortIdMap;
 import org.apache.beam.sdk.fn.data.BeamFnDataInboundObserver2;
 import org.apache.beam.sdk.fn.data.BeamFnDataOutboundAggregator;
@@ -156,6 +155,7 @@ public class ProcessBundleHandler {
   private final FinalizeBundleHandler finalizeBundleHandler;
   private final ShortIdMap shortIds;
   private final boolean runnerAcceptsShortIds;
+  private final ExecutionStateSampler executionStateSampler;
   private final Map<String, PTransformRunnerFactory> urnToPTransformRunnerFactoryMap;
   private final PTransformRunnerFactory defaultPTransformRunnerFactory;
   private final Cache<Object, Object> processWideCache;
@@ -170,6 +170,7 @@ public class ProcessBundleHandler {
       BeamFnStateGrpcClientCache beamFnStateGrpcClientCache,
       FinalizeBundleHandler finalizeBundleHandler,
       ShortIdMap shortIds,
+      ExecutionStateSampler executionStateSampler,
       Cache<Object, Object> processWideCache) {
     this(
         options,
@@ -179,6 +180,7 @@ public class ProcessBundleHandler {
         beamFnStateGrpcClientCache,
         finalizeBundleHandler,
         shortIds,
+        executionStateSampler,
         REGISTERED_RUNNER_FACTORIES,
         processWideCache,
         new BundleProcessorCache());
@@ -193,6 +195,7 @@ public class ProcessBundleHandler {
       BeamFnStateGrpcClientCache beamFnStateGrpcClientCache,
       FinalizeBundleHandler finalizeBundleHandler,
       ShortIdMap shortIds,
+      ExecutionStateSampler executionStateSampler,
       Map<String, PTransformRunnerFactory> urnToPTransformRunnerFactoryMap,
       Cache<Object, Object> processWideCache,
       BundleProcessorCache bundleProcessorCache) {
@@ -206,6 +209,7 @@ public class ProcessBundleHandler {
     this.runnerAcceptsShortIds =
         runnerCapabilities.contains(
             BeamUrns.getUrn(RunnerApi.StandardRunnerProtocols.Enum.MONITORING_INFO_SHORT_IDS));
+    this.executionStateSampler = executionStateSampler;
     this.urnToPTransformRunnerFactoryMap = urnToPTransformRunnerFactoryMap;
     this.defaultPTransformRunnerFactory =
         new UnknownPTransformRunnerFactory(urnToPTransformRunnerFactoryMap.keySet());
@@ -366,7 +370,8 @@ public class ProcessBundleHandler {
                     @Override
                     public <T> void addPCollectionConsumer(
                         String pCollectionId, FnDataReceiver<WindowedValue<T>> consumer) {
-                      pCollectionConsumerRegistry.register(pCollectionId, pTransformId, consumer);
+                      pCollectionConsumerRegistry.register(
+                          pCollectionId, pTransformId, pTransform.getUniqueName(), consumer);
                     }
 
                     @Override
@@ -420,12 +425,14 @@ public class ProcessBundleHandler {
 
                     @Override
                     public void addStartBundleFunction(ThrowingRunnable startFunction) {
-                      startFunctionRegistry.register(pTransformId, startFunction);
+                      startFunctionRegistry.register(
+                          pTransformId, pTransform.getUniqueName(), startFunction);
                     }
 
                     @Override
                     public void addFinishBundleFunction(ThrowingRunnable finishFunction) {
-                      finishFunctionRegistry.register(pTransformId, finishFunction);
+                      finishFunctionRegistry.register(
+                          pTransformId, pTransform.getUniqueName(), finishFunction);
                     }
 
                     @Override
@@ -506,7 +513,8 @@ public class ProcessBundleHandler {
       ExecutionStateTracker stateTracker = bundleProcessor.getStateTracker();
 
       try (HandleStateCallsForBundle beamFnStateClient = bundleProcessor.getBeamFnStateClient()) {
-        try (Closeable closeTracker = stateTracker.activate()) {
+        stateTracker.start(request.getInstructionId());
+        try {
           // Already in reverse topological order so we don't need to do anything.
           for (ThrowingRunnable startFunction : startFunctionRegistry.getFunctions()) {
             LOG.debug("Starting function {}", startFunction);
@@ -542,30 +550,35 @@ public class ProcessBundleHandler {
             LOG.debug("Finishing function {}", finishFunction);
             finishFunction.run();
           }
-        }
 
-        // If bundleProcessor has not flushed any elements, embed them in response.
-        embedOutboundElementsIfApplicable(response, bundleProcessor);
+          // If bundleProcessor has not flushed any elements, embed them in response.
+          embedOutboundElementsIfApplicable(response, bundleProcessor);
 
-        // Add all checkpointed residuals to the response.
-        response.addAllResidualRoots(bundleProcessor.getSplitListener().getResidualRoots());
+          // Add all checkpointed residuals to the response.
+          response.addAllResidualRoots(bundleProcessor.getSplitListener().getResidualRoots());
 
-        // Add all metrics to the response.
-        Map<String, ByteString> monitoringData = finalMonitoringData(bundleProcessor);
-        if (runnerAcceptsShortIds) {
-          response.putAllMonitoringData(monitoringData);
-        } else {
-          for (Map.Entry<String, ByteString> metric : monitoringData.entrySet()) {
-            response.addMonitoringInfos(
-                shortIds.get(metric.getKey()).toBuilder().setPayload(metric.getValue()));
+          // Add all metrics to the response.
+          bundleProcessor.getProgressRequestLock().lock();
+          Map<String, ByteString> monitoringData = finalMonitoringData(bundleProcessor);
+          if (runnerAcceptsShortIds) {
+            response.putAllMonitoringData(monitoringData);
+          } else {
+            for (Map.Entry<String, ByteString> metric : monitoringData.entrySet()) {
+              response.addMonitoringInfos(
+                  shortIds.get(metric.getKey()).toBuilder().setPayload(metric.getValue()));
+            }
           }
-        }
 
-        if (!bundleProcessor.getBundleFinalizationCallbackRegistrations().isEmpty()) {
-          finalizeBundleHandler.registerCallbacks(
-              bundleProcessor.getInstructionId(),
-              ImmutableList.copyOf(bundleProcessor.getBundleFinalizationCallbackRegistrations()));
-          response.setRequiresFinalization(true);
+          if (!bundleProcessor.getBundleFinalizationCallbackRegistrations().isEmpty()) {
+            finalizeBundleHandler.registerCallbacks(
+                bundleProcessor.getInstructionId(),
+                ImmutableList.copyOf(bundleProcessor.getBundleFinalizationCallbackRegistrations()));
+            response.setRequiresFinalization(true);
+          }
+        } finally {
+          // We specifically deactivate state tracking while we are holding the progress request and
+          // sampling locks.
+          stateTracker.reset();
         }
       }
 
@@ -668,15 +681,6 @@ public class ProcessBundleHandler {
   private Map<String, ByteString> intermediateMonitoringData(BundleProcessor bundleProcessor)
       throws Exception {
     Map<String, ByteString> monitoringData = new HashMap<>();
-    // Get start bundle Execution Time Metrics.
-    monitoringData.putAll(
-        bundleProcessor.getStartFunctionRegistry().getExecutionTimeMonitoringData(shortIds));
-    // Get process bundle Execution Time Metrics.
-    monitoringData.putAll(
-        bundleProcessor.getpCollectionConsumerRegistry().getExecutionTimeMonitoringData(shortIds));
-    // Get finish bundle Execution Time Metrics.
-    monitoringData.putAll(
-        bundleProcessor.getFinishFunctionRegistry().getExecutionTimeMonitoringData(shortIds));
     // Extract MonitoringInfos that come from the metrics container registry.
     monitoringData.putAll(
         bundleProcessor.getMetricsContainerRegistry().getMonitoringData(shortIds));
@@ -689,17 +693,7 @@ public class ProcessBundleHandler {
 
   private Map<String, ByteString> finalMonitoringData(BundleProcessor bundleProcessor)
       throws Exception {
-    bundleProcessor.getProgressRequestLock().lock();
     HashMap<String, ByteString> monitoringData = new HashMap<>();
-    // Get start bundle Execution Time Metrics.
-    monitoringData.putAll(
-        bundleProcessor.getStartFunctionRegistry().getExecutionTimeMonitoringData(shortIds));
-    // Get process bundle Execution Time Metrics.
-    monitoringData.putAll(
-        bundleProcessor.getpCollectionConsumerRegistry().getExecutionTimeMonitoringData(shortIds));
-    // Get finish bundle Execution Time Metrics.
-    monitoringData.putAll(
-        bundleProcessor.getFinishFunctionRegistry().getExecutionTimeMonitoringData(shortIds));
     // Extract MonitoringInfos that come from the metrics container registry.
     monitoringData.putAll(
         bundleProcessor.getMetricsContainerRegistry().getMonitoringData(shortIds));
@@ -744,8 +738,8 @@ public class ProcessBundleHandler {
     BundleProgressReporter.InMemory bundleProgressReporterAndRegistrar =
         new BundleProgressReporter.InMemory();
     MetricsContainerStepMap metricsContainerRegistry = new MetricsContainerStepMap();
-    ExecutionStateTracker stateTracker =
-        new ExecutionStateTracker(ExecutionStateSampler.instance());
+    ExecutionStateTracker stateTracker = executionStateSampler.create();
+    bundleProgressReporterAndRegistrar.register(stateTracker);
     PCollectionConsumerRegistry pCollectionConsumerRegistry =
         new PCollectionConsumerRegistry(
             metricsContainerRegistry,
@@ -757,10 +751,10 @@ public class ProcessBundleHandler {
 
     PTransformFunctionRegistry startFunctionRegistry =
         new PTransformFunctionRegistry(
-            metricsContainerRegistry, stateTracker, ExecutionStateTracker.START_STATE_NAME);
+            metricsContainerRegistry, shortIds, stateTracker, Urns.START_BUNDLE_MSECS);
     PTransformFunctionRegistry finishFunctionRegistry =
         new PTransformFunctionRegistry(
-            metricsContainerRegistry, stateTracker, ExecutionStateTracker.FINISH_STATE_NAME);
+            metricsContainerRegistry, shortIds, stateTracker, Urns.FINISH_BUNDLE_MSECS);
     List<ThrowingRunnable> resetFunctions = new ArrayList<>();
     List<ThrowingRunnable> tearDownFunctions = new ArrayList<>();
 
@@ -1118,10 +1112,7 @@ public class ProcessBundleHandler {
           this.bundleCache = null;
         }
       }
-      getStartFunctionRegistry().reset();
-      getFinishFunctionRegistry().reset();
       getSplitListener().clear();
-      getpCollectionConsumerRegistry().reset();
       getMetricsContainerRegistry().reset();
       getStateTracker().reset();
       getBundleFinalizationCallbackRegistrations().clear();