Msec counters benchmarks, 2.5x-3.5x faster, larger bundles have bigger wins

```
Benchmark                                                               Mode  Cnt         Score        Error  Units
ExecutionStateSamplerBenchmark.testLargeBundleHarnessStateSampler      thrpt   25     74093.492 ±    364.322  ops/s
ExecutionStateSamplerBenchmark.testLargeBundleRunnersCoreStateSampler  thrpt   25     19838.512 ±     81.986  ops/s
ExecutionStateSamplerBenchmark.testTinyBundleHarnessStateSampler       thrpt   25  10187284.157 ± 128030.350  ops/s
ExecutionStateSamplerBenchmark.testTinyBundleRunnersCoreStateSampler   thrpt   25   4099369.479 ±  25219.934  ops/s
```

Before
```
Benchmark                               Mode   Cnt      Score     Error  Units
ProcessBundleBenchmark.testTinyBundle   thrpt   25  28624.904 ± 231.755  ops/s
ProcessBundleBenchmark.testLargeBundle  thrpt   25   1128.407 ±  22.196  ops/s
```

After
```
Benchmark                               Mode   Cnt      Score     Error  Units
ProcessBundleBenchmark.testTinyBundle   thrpt   25  29091.576 ± 103.945  ops/s
ProcessBundleBenchmark.testLargeBundle  thrpt   25   1158.686 ±   6.835  ops/s
```

------------------------

Thank you for your contribution! Follow this checklist to help us incorporate your contribution quickly and easily:

 - [ ] [**Choose reviewer(s)**](https://beam.apache.org/contribute/#make-your-change) and mention them in a comment (`R: @username`).
 - [ ] Mention the appropriate issue in your description (for example: `addresses #123`), if applicable. This will automatically add a link to the pull request in the issue. If you would like the issue to automatically close on merging the pull request, comment `fixes #<ISSUE NUMBER>` instead.
 - [ ] Update `CHANGES.md` with noteworthy changes.
 - [ ] If this contribution is large, please file an Apache [Individual Contributor License Agreement](https://www.apache.org/licenses/icla.pdf).

See the [Contributor Guide](https://beam.apache.org/contribute) for more tips on [how to make review process smoother](https://beam.apache.org/contribute/#make-reviewers-job-easier).

To check the build health, please visit [https://github.com/apache/beam/blob/master/.test-infra/BUILD_STATUS.md](https://github.com/apache/beam/blob/master/.test-infra/BUILD_STATUS.md)

GitHub Actions Tests Status (on master branch)
------------------------------------------------------------------------------------------------
[![Build python source distribution and wheels](https://github.com/apache/beam/workflows/Build%20python%20source%20distribution%20and%20wheels/badge.svg?branch=master&event=schedule)](https://github.com/apache/beam/actions?query=workflow%3A%22Build+python+source+distribution+and+wheels%22+branch%3Amaster+event%3Aschedule)
[![Python tests](https://github.com/apache/beam/workflows/Python%20tests/badge.svg?branch=master&event=schedule)](https://github.com/apache/beam/actions?query=workflow%3A%22Python+Tests%22+branch%3Amaster+event%3Aschedule)
[![Java tests](https://github.com/apache/beam/workflows/Java%20Tests/badge.svg?branch=master&event=schedule)](https://github.com/apache/beam/actions?query=workflow%3A%22Java+Tests%22+branch%3Amaster+event%3Aschedule)

See [CI.md](https://github.com/apache/beam/blob/master/CI.md) for more information about GitHub Actions CI.