@lukecwik 2022-06-29
Tiny bundle:
![large bundle](https://user-images.githubusercontent.com/10078956/176501848-df2c04b1-7d78-4d3e-95b6-2fd3e08754b4.png)

Large bundle:
![tiny bundle](https://user-images.githubusercontent.com/10078956/176501856-1ee0a341-a44c-4ff4-b09a-51fe2d43acd8.png)