@lukecwik 2022-07-08
> > The overhead that I see isn't the sampler but the cost of producing an output and sending it to the next fused transform.
> > For trivial DoFns the overhead can be as high as 50% of the CPU time for processing an element.
> 
> Oh yeah totally, I think this is great stuff, sorry I didn't mean to imply otherwise.

Never thought you did, was giving context.